Moved some app logic from PostDAO to PostManager
